+++ src/main/kotlin/com/kennycason/nn/data/MatrixImageEncoder.kt
@@ -1,7 +1,7 @@
 package com.kennycason.nn.data
 
-import org.jblas.DoubleMatrix
+import org.jblas.FloatMatrix
 
 interface MatrixImageEncoder {
-    fun encode(image: Image): DoubleMatrix
+    fun encode(image: Image): FloatMatrix
 }
