+++ src/main/kotlin/com/kennycason/nn/data/MatrixNBitImageEncoder.kt
@@ -1,6 +1,6 @@
 package com.kennycason.nn.data
 
-import org.jblas.DoubleMatrix
+import org.jblas.FloatMatrix
 import java.awt.image.BufferedImage
 
 
@@ -9,8 +9,8 @@ class MatrixNBitImageEncoder(private val bits: Int = 24) : MatrixImageEncoder {
 
     private val HIGH_BIT_FLAG = 8388608
 
-    override fun encode(image: Image): DoubleMatrix {
-        val data = DoubleMatrix(1, image.width() * bits * image.height())
+    override fun encode(image: Image): FloatMatrix {
+        val data = FloatMatrix(1, image.width() * bits * image.height())
         val bi = image.data()
         for (y in 0 until image.height()) {
             for (x in 0 until image.width()) {
@@ -20,14 +20,14 @@ class MatrixNBitImageEncoder(private val bits: Int = 24) : MatrixImageEncoder {
         return data
     }
 
-    private fun read(data: DoubleMatrix, bi: BufferedImage, x: Int, y: Int) {
+    private fun read(data: FloatMatrix, bi: BufferedImage, x: Int, y: Int) {
         var flag = HIGH_BIT_FLAG
         var offset = 0
         while (flag > 0) {
             val rgb = bi.getRGB(x, y) and 0xFFFFFF
             val set = rgb and flag == flag
             val index = y * bi.width * bits + (x * bits + offset)
-            data.data[index] = if (set) 1.0 else 0.0
+            data.data[index] = if (set) 1.0f else 0.0f
             offset++
             flag = flag shr RGB_BITS / bits
         }
