+++ src/main/kotlin/com/kennycason/nn/learning_rate/LearningRate.kt
@@ -2,5 +2,5 @@ package com.kennycason.nn.learning_rate
 
 
 interface LearningRate {
-    fun apply(): Double
+    fun apply(): Float
 }
